fix sort problem with sub sections

+++ frontend/models/models/RecipeView.php
@@ -18,8 +18,8 @@ class RecipeView extends RecipeNutrient
 
     public function sort($array) {
         usort($array, static function(self $a, self $b) {
-            return ((int)($a->section->name) > (int)($b->section->name)) || (
-                    (int)($a->section->name) === (int)($b->section->name) && ($a->catalog && $b->catalog && $a->catalog->id > $b->catalog->id)
+            return ((float)($a->section->name) > (float)($b->section->name)) || (
+                    (float)($a->section->name) === (float)($b->section->name) && ($a->catalog && $b->catalog && $a->catalog->id > $b->catalog->id)
                 );
         });
         return $array;
